// packages/codplay-v1/src/animation/types.ts
export type AnimatedProperty = string
// ...
export type TransitionEase = string | {
  type: 'physics'
  velocity?: number
  mass?: number
  stiffness?: number
  damping?: number
  bounce?: number
}

export type AnimationTimerOptions = {
  duration?: number
  delayMs?: number
  loopDelayMs?: number
  reversed?: boolean
  alternate?: boolean
  loop?: boolean | number
  ease?: TransitionEase
  stagger?: number
  ignoreDuration?: boolean
}
// ...
export type TransitionRequest = {
  transitionId: string
  eventId: string
  eventName: string
  listenerId: string
  property: AnimatedProperty
  target: unknown
  from?: number | string
  to: number | string
  finalValue?: number | string
  modifier?: (value: number) => number | string
} & Required<Pick<AnimationTimerOptions, 'duration'>> & {
  easing?: string
} & AnimationTimerOptions & {
  composition?: 'merge' | 'replace'
  cleanupStyleProperty?: 'width' | 'height'
  onFinalize?: (reason: 'completed' | 'stopped') => void
  onFrame?: () => void
  group?: {
    id: string
    total: number
    onGroupFinalize: (reason: 'completed' | 'stopped') => void
  }
}